requirements.txt

flask
google-generativeai
pillow
gunicorn
